+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 6.1.4 (TBD)
 
-* TBD
+* geopackage-android version 6.1.4
+* FeatureOverlayQuery tileBounds methods
 
 ## [6.1.3](https://github.com/ngageoint/geopackage-android-map/releases/tag/6.1.3) (12-17-2021)